v4.6.0.  Fixes for 64-bit systems, code/directory reorg, more Makefiles.

+++ src/vm/globs.h
@@ -6,10 +6,11 @@
 #define GLOBS_H
 #include <sys/types.h>
 #include <stdio.h>
+#include <stdint.h>
 
 extern int fileIn(FILE * fp), fileOut(FILE * fp);
 extern void sysError(char *);
-extern void sysErrorInt(char *, int);
+extern void sysErrorInt(char *, intptr_t);
 extern void sysErrorStr(char *, char *);
 extern void flushCache(void);
 extern struct object *primitive(int, struct object *, int *);
